# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Lost & Found API
  description: API for managing lost items and claims in the Lost & Found system.
  version: 1.0.0
paths:
  /:
    get:
      summary: Get home page with all lost items
      operationId: getHomePage
      responses:
        '200':
          description: HTML page displaying all lost items
        '500':
          description: Internal server error

  /admin/upload:
    get:
      summary: Upload lost items page
      operationId: getUploadPage
      responses:
        '302':
          description: Redirect to upload HTML page

    post:
      summary: Upload lost items from a file
      operationId: uploadLostItems
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: File uploaded successfully
        '500':
          description: Internal server error

  /items:
    get:
      summary: Get all lost items
      operationId: getAllLostItems
      responses:
        '200':
          description: HTML page displaying all lost items
        '500':
          description: Internal server error

  /items/{id}:
    get:
      summary: Get lost item by ID
      operationId: getLostItemById
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the lost item
          schema:
            type: integer
      responses:
        '200':
          description: HTML page displaying the lost item's details
        '404':
          description: Lost item not found

  /claims/{id}:
    get:
      summary: Get claim by ID
      operationId: getClaimById
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the claim
          schema:
            type: integer
      responses:
        '200':
          description: HTML page displaying the claim's details
        '404':
          description: Claim not found

  /claims/{lostItemId}/claim:
    post:
      summary: Claim a lost item
      operationId: claimLostItem
      parameters:
        - name: lostItemId
          in: path
          required: true
          description: ID of the lost item to claim
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                quantityClaimed:
                  type: integer
      responses:
        '302':
          description: Redirect to items page
        '400':
          description: Invalid input

  /claims:
    get:
      summary: Get all claims
      operationId: getAllClaims
      responses:
        '200':
          description: HTML page displaying all claims
        '500':
          description: Internal server error
components:
  schemas:
    LostItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        dateLost:
          type: string
          format: date
    Claim:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        lostItemId:
          type: integer
        quantityClaimed:
          type: integer
        status:
          type: string